Upgrade xorg-sgml-doctools to Autoconf 2.60 directory architecture
Currently the value of sgmlrootdir is hard coded relative to $prefix.
In Autoconf 2.60, $datarootdir has been added to define the architecture
independent data directory.

  --datarootdir=DIR       read-only arch.-independent data root [PREFIX/share]

Relative to that location, a number of subdirectories have been defined:

  --datadir=DIR      read-only architecture-independent data [DATAROOTDIR]
  --infodir=DIR      info documentation [DATAROOTDIR/info]
  --localedir=DIR    locale-dependent data [DATAROOTDIR/locale]
  --mandir=DIR       man documentation [DATAROOTDIR/man]
  --docdir=DIR       documentation root
                     [DATAROOTDIR/doc/xorg-sgml-doctools]
  --htmldir=DIR      html documentation [DOCDIR]
  --dvidir=DIR       dvi documentation [DOCDIR]
  --pdfdir=DIR       pdf documentation [DOCDIR]
  --psdir=DIR        ps documentation [DOCDIR]

The sgmlrootdir should hang off datarootdir. If there is a need to specify
a different location, a new configure option should be added:

  --sgmldir=DIR      sgml stylesheets and entities [DATAROOTDIR/sgml]

An sgmlrootdir Automake variable is defined in configure.ac to provide
a unique value holder for the sgmlrootdir.

Variable PACKAGE_VERSION is preferred over the undocumented VERSION variable.

Signed-off-by: Gaetan Nadon <memsize@videotron.ca>