DocBook/XML: add support for docbook external references
Documents external references
-----------------------------
A new feature with version 1.7 is references to documents in other
packages. For example a protocol specifications may refer to another
one or to a library implementing it.

This feature works with PDF, HTML, PS (and epub?) generated documents
and will generate links to the other versions of the same type, i.e. html
docs will link to other html docs,  and pdf docs will link to other pdf
docs.

New: generating target databases
--------------------------------
All documents will generate a target db, analogous to a symbol table
which will be consulted by other documents to find linking information.
This is accomplished with xsltproc.

New: consulting target databases
--------------------------------
While documents are being converted from xml to html/pdf/ps,
they will consult the target dbs of each document they which to refer to.
This is accomplished with xmlto.

New: master database
--------------------
A hand written database has been created which provides an association
between the target dbs and the document themselves. It also reflects
the hierarchical structure of the documents relative to the installation dir.

New: xorg-xhtml.xsl and xorg-fo.xsl
-----------------------------------
A number of motivations are behind the new (and split) versions of xorg.xsl.
This module must remain backward compatible to build previous versions
of documentation. It was originally designed a while ago for sgml.
Given that all documents makefile had to change, it seemed a good opportunity
to break from the past. Among the differences:

The location of xorg.css is no longer set in the stylesheet.
The chunked html is no longer supported.
The html and fo parameters are kept separate.
Additional parameters for external references.
Now allows different values for html/fo xsl for the same parameter
should that be required.
The xorg-xhtml.xsl imports html/docbook.xsl
The xorg-fo.xsl imports fo/docbook.xsl

Co-authored-by: Matt Dew <marcoz@osource.org>

Acked-by: Dan Nicholson <dbn.lists@gmail.com>
Signed-off-by: Gaetan Nadon <memsize@videotron.ca>