Masterdb: read the target dbs from the usual doc directory
This is a coordinated change with the packages building DocBook/XML.
The target dbs are now installed alongside the generated documents.
Currently they are installed in $prefix/sgml/X11/dbs alongside masterdb which
has the potential of installing outside the package prefix and cause
distcheck to fail when user does not have write permission in this package.

The master db xi:include statement is changed. It uses a fully qualified
path but could also have used a relative path although it is difficult to
ensure it would work reliably. The generated html/pdf link are still relative.

The README elaborates on requirements which previously existed but were not
introduced by this patch.

Signed-off-by: Gaetan Nadon <memsize@videotron.ca>