Review documents cascading style sheet
There is no drastic changes in style, it will go unnoticed
for the most part. Many inconsistencies in titles and headings
were fixed in both books and articles.

The large libX11 specs was used a reference. The border above
each and every section was removed, keeping only level 2 headings.
It alleviates the sense of "being lost". An additional step might be
the section numbering.

The body background and foreground colors are no longer set, which
will not interfere with the user own styling. This should fix #37535.

The figure-float was removed as it was specific for one figure and
did not yield nice results. There is virtually no room for
improvements with what docbook has to offer.

All xorg docs were reviewed to ensure nothing got badly broken.
A few changes in the doc source was made in libX11 and similar
changes need to be done in the remaining xorg docs. They only affect
minor spacing issues.

Signed-off-by: Gaetan Nadon <memsize@videotron.ca>