masterdb: add olink support for chunked html
For olinks to chunked html, the masterdb cannot contain the baseuri
attribute. The xsl parameter olink.base.uri needs to be used in makefiles
for non chunked html while the chunked html does not use it.

The txt and ps documents cannot be an end point of olinks, so no target
dbs will be generated for those formats, hence no masterdb needed either.

Signed-off-by: Gaetan Nadon <memsize@videotron.ca>