CSS: authors and affiliation spacing
The objective is to have more spacing between each author so their affiliation
is visually grouped with its author. The previous method was not reliable as
it was making an assumption about the preceeding element.

The new method of reducing the bottom margin for the author should be more
reliable.

Signed-off-by: Gaetan Nadon <memsize@videotron.ca>