Embed CSS styles inside the HTML HEAD element
Rather than referring to the external xorg.css stylesheet, embed the content
of the file in the html output produced. This is accomplished by adding
an xsl:template to the customization layer. The content of xorg.css must not
be encoded due to the '<' character.

This makes the whole html docs tree much more relocatable.
In addition, it eliminates xorg.css as a runtime file which makes
xorg-sgml-doctools a build time only package.

Signed-off-by: Gaetan Nadon <memsize@videotron.ca>