CSS style: support styling for multi licensed docs
For documentation having multiple licenses, the copyright and legalnotice
elements sequence cannot instantiated multiple times.
The copyright notice and license text are therefore coded inside a legalnotice
element. The role attribute on the paragraph is used to allow styling of the
copyright notice text which should not be italicized.
Also restore the font size to the default "medium" size.

Markup example:
<para role="multiLicensing">Copyright © 1994 X Consortium</para>

Signed-off-by: Gaetan Nadon <memsize@videotron.ca>